Performance branch (#1)
# Performance Branch

The most computationally expensive thing to do on any computer is render a font. It's what Babbage fought with for years.

I profiled the main loop of this code by vendoring in Luma, removing the threadpool, and running cProfile over it. That revealed two things:

* Running the seconds as "hotspot" was burning up the CPU
* The rest of the CPU time was spent "scrolling" the "calling at stations" as it required a full re-render of the string each time a character was dropped off the front of the string.
* We've changed the scrolling behaviour to scroll the bitmap at 1 pixel per frame, rather than 1 character per frame. Smooth!
  * There's also a snazzy little rising-up animation for fun.
* The other frequent font calls (in the loop) have all been bitmap cached too.
* The frame regulator is now configurable; it fights with the CPU on a Pi Zero and is better disabled there; but on a Pi3 you want a regulator to stop burning the CPU up unnecessarily.

So, main fixes were to put the seconds as an interval-updated zone with 0.1 second resolution, and to pre-render all other commonly used TTF operations in the main loop.

## Results

On a Pi Zero (the oldest possible device!), here's the performance on "main":

![353895428_652705433582787_8305305711463697195_n](https://github.com/CalamityJames/train-departure-display/assets/1850718/823cfcc8-1f6b-4730-ae5d-f49e655af10f)

And here's the performance on "performance", with `targetFPS` set to `0`:

![image](https://github.com/CalamityJames/train-departure-display/assets/1850718/c1a260a1-cd26-4872-b204-4654293caa9a)

# Changelog

See updated CHANGELOG

---------

Co-authored-by: James <me@calamityjames.dev>